Found a security problem with Windows, where % in mesages can be used to get environment variables.
git-svn-id: https://optime.repos.se/svn/code/repos/trunk@1730 94f62164-8fde-0310-83d7-e1105c8c1596

+++ www/open/log/index.php
@@ -4,26 +4,25 @@ require_once(dirname(dirname(dirname(__FILE__))) . "/account/login.inc.php" );
 
 define('STYLESHEET','../../view/log.xsl');
 
-$url = urlEncodeNames(getTargetUrl());
-if (!getPath()) {
-	echo("Log withoug the 'path' parameter has not been implemented yet");exit;
+$url = getTargetUrl();
+if (isFile($url)) {
+	trigger_error("History for a single file has not been implemented yet");
+	exit;
 }
 
-$cmd = "log -v --xml --incremental ".$url;
+$cmd = 'log -v --xml --incremental '.escapeArgument($url); // escapeArgument not needed, because it is URLencoded 
 
 // start output
 
 // Set HTTP output character encoding to SJIS
-mb_http_output('UTF-8');
-// Start buffering and specify "mb_output_handler" as callback function
-ob_start('mb_output_handler');
+//should be configured in server //mb_http_output('UTF-8');
 
 // passthrough with stylesheet
 header('Content-type: text/xml');
 echo '<?xml version="1.0" encoding="utf-8"?>' . "\n";
 echo '<?xml-stylesheet type="text/xsl" href="' . STYLESHEET . '"?>' . "\n";
 echo "<!-- SVN log for $url -->\n";
-echo '<log repo="'.getRepositoryUrl().'" path="'.getPath().'">' . "\n";
+echo '<log repo="'.getRepositoryUrl().'" path="'.getTarget().'">' . "\n";
 $returnvalue = login_svnPassthru($cmd);
 if ($returnvalue) login_handleSvnError($cmd, $returnvalue);
 echo '</log>';
